requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
groq==0.9.0
httpx==0.27.2
pytesseract==0.3.13
Pillow==10.4.0
chromadb==0.5.5
sentence-transformers==3.1.1
PyMuPDF==1.24.10
reportlab==4.2.2
aiofiles==24.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
